Introduce a unified CLI entrypoint to centralize command parsing
